arc_solver/requirements.txt
numpy>=1.21.0
pydantic>=2.0.0
tenacity>=8.2.0
anthropic>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
